+++ CHANGELOG.md
@@ -25,6 +25,7 @@ If there are not many items, just list them sequentially.
 * Use new bot profile UI for 'explore near by' screen (#402)
 * Clean up debugging code. Improve logging and error handling. 
 * Bot Profile: Remove Hamburger and Message Icon (#404)
+* Fixed: Bot Creation: Not able to add Note + Photo during Bot Creation Flow (#388)
 * Set image maximum upload size to 3000x3000 (#386)